+++ Sources/Client/Model/Filter.swift
@@ -24,6 +24,11 @@ import Foundation
 public enum Filter: Encodable, CustomStringConvertible {
     public typealias Key = String
     
+    /// No filter.
+    /// Can be used to get all users in queryUsers
+    /// Warning: Should not be used in queryChannels.
+    case none
+    
     // MARK: Operators
     
     /// An equal operator.
@@ -58,6 +63,8 @@ public enum Filter: Encodable, CustomStringConvertible {
     
     public var description: String {
         switch self {
+        case .none:
+            return ""
         case let .equal(key, object):
             return "\(key) = \(object)"
         case let .notEqual(key, object):
@@ -94,6 +101,8 @@ public enum Filter: Encodable, CustomStringConvertible {
         var operands: [Encodable] = []
         
         switch self {
+        case .none:
+            return
         case let .equal(key, object):
             try [key: AnyEncodable(object)].encode(to: encoder)
             return